requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
colorzero==1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-cors-headers==2.5.2
django-json-response==1.1.5
django-progressive-web-app==0.1.1
gpiozero==1.5.0
gunicorn==19.9.0
more-itertools==7.0.0
paho-mqtt==1.4.0
pathlib2==2.3.3
picamera==1.13
pigpio==1.42
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyserial==3.4
pytest==4.4.0
python-decouple==3.1
pytz==2018.9
RPi.GPIO==0.6.5
RPIO==0.10.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0

